Added interface for NATS Core and Streaming functionality
